+++ babel.config.js
@@ -1,3 +1,6 @@
+const path = require('path');
+const glob = require('glob');
+
 module.exports = {
   presets: [
     // Polyfills
@@ -13,6 +16,34 @@ module.exports = {
     // Devs tend to write `import { someIcon } from '@patternfly/react-icons';`
     // This transforms the import to be specific which prevents having to parse 2k+ icons
     // Also prevents potential bundle size blowups with CJS
+    [
+      'transform-imports',
+      {
+        '@patternfly/react-core': {
+          transform: (importName) => {
+            let res;
+            const files = glob.sync(
+              path.resolve(
+                __dirname,
+                `./node_modules/@patternfly/react-core/dist/js/**/${importName}.js`
+              )
+            );
+            if (files.length > 0) {
+              res = files[0];
+            } else {
+              throw `File with importName ${importName} does not exist`;
+            }
+
+            res = res.replace(path.resolve(__dirname, './node_modules/'), '');
+            res = res.replace(/^\//, '');
+            return res;
+          },
+          preventFullImport: false,
+          skipDefaultConversion: true,
+        },
+      },
+      'react-core',
+    ],
     [
       'transform-imports',
       {